Bug 641718 - Backout 44d43f095a4f. r=backout
UltraBlame original commit: c1553501c4966df32d22dcd117532d62bdfe9731

+++ content/media/wave/nsWaveReader.h
@@ -77,12 +77,12 @@ private:
   
   
   
-  double BytesToTime(PRInt64 aBytes) const;
+  float BytesToTime(PRInt64 aBytes) const;
 
   
   
   
-  PRInt64 TimeToBytes(double aTime) const;
+  PRInt64 TimeToBytes(float aTime) const;
